+++ packages/dynamics-lib/src/entities/__tests__/contact.entity.spec.js
@@ -25,7 +25,8 @@ describe('contact entity', () => {
         defra_country: 910400184,
         defra_preferredmethodofconfirmation: 910400002,
         defra_preferredmethodofnewsletter: 910400000,
-        defra_preferredmethodofreminder: 910400001
+        defra_preferredmethodofreminder: 910400001,
+        defra_postalfulfilment: true
       },
       optionSetData
     )
@@ -46,7 +47,8 @@ describe('contact entity', () => {
       country: expect.objectContaining({ id: 910400184, label: 'United Kingdom', description: 'GB' }),
       preferredMethodOfConfirmation: expect.objectContaining({ id: 910400002, label: 'Text', description: 'Text' }),
       preferredMethodOfNewsletter: expect.objectContaining({ id: 910400000, label: 'Email', description: 'Email' }),
-      preferredMethodOfReminder: expect.objectContaining({ id: 910400001, label: 'Letter', description: 'Letter' })
+      preferredMethodOfReminder: expect.objectContaining({ id: 910400001, label: 'Letter', description: 'Letter' }),
+      postalFulfilment: true
     }
 
     expect(contact).toBeInstanceOf(Contact)
@@ -72,6 +74,7 @@ describe('contact entity', () => {
     contact.preferredMethodOfConfirmation = optionSetData.defra_preferredcontactmethod.options['910400002']
     contact.preferredMethodOfNewsletter = optionSetData.defra_preferredcontactmethod.options['910400000']
     contact.preferredMethodOfReminder = optionSetData.defra_preferredcontactmethod.options['910400001']
+    contact.postalFulfilment = true
 
     const dynamicsEntity = contact.toRequestBody()
     expect(dynamicsEntity).toMatchObject(
@@ -90,7 +93,8 @@ describe('contact entity', () => {
         defra_country: 910400184,
         defra_preferredmethodofconfirmation: 910400002,
         defra_preferredmethodofnewsletter: 910400000,
-        defra_preferredmethodofreminder: 910400001
+        defra_preferredmethodofreminder: 910400001,
+        defra_postalfulfilment: true
       })
     )
   })