+++ packages/gafl-webapp-service/src/__tests__/server-ga-integration.spec.js
@@ -27,6 +27,9 @@ jest.mock('../constants', () => ({
   },
   CommonResults: {
     OK: 'okay'
+  },
+  allowsPhysicalLicence: {
+    YES: 'yes'
   }
 }))