+++ packages/sales-api-service/src/services/transactions/__tests__/process-transaction-queue.spec.js
@@ -59,6 +59,7 @@ describe('transaction service', () => {
   beforeAll(() => {
     TRANSACTION_STAGING_TABLE.TableName = 'TestTable'
   })
+
   beforeEach(AwsMock.__resetAll)
 
   describe('processQueue', () => {